+++ packages/iti-react-core/src/_Util/index.ts
@@ -1,6 +1,5 @@
 ﻿export * from './ConversionFunctions'
 export * from './PaginationUtil'
-export * from './FormatUrlParams'
 export * from './GetGuid'
 export * from './DateTimeConversion'
 export * from './PhoneNumberUtil'